requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.2.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.1
black==22.10.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
constantly==15.1.0
contourpy==1.0.7
cryptography==40.0.1
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.0.4
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.0
fastjsonschema==2.16.3
filelock==3.10.7
flake8==6.0.0
flake8-annotations==2.9.1
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.0.6
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==2.0.5
multidict==6.0.4
mypy-extensions==0.4.3
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
ngrok==0.2.2
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandocfilters==1.5.0
parsel==1.7.0
parso==0.8.3
pathspec==0.10.1
pep8-naming==0.13.2
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
Protego==0.2.1
psutil==5.9.4
psycopg2-binary==2.9.5
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.7
PyDispatcher==2.0.7
pyflakes==2.4.0
Pygments==2.15.0
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.6
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Scrapy==2.8.0
selenium==4.8.3
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.26.1
terminado==0.17.1
tinycss2==1.2.1
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.5.0
tzdata==2022.7
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
w3lib==2.1.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
wsproto==1.2.0
xlwt==1.3.0
yarl==1.9.2
zope.interface==6.0
